+++ composer.json
@@ -20,18 +20,14 @@
         }
     ],
     "require": {
-        "php": "^8.1"
+        "php": "^8.1",
+        "jenky/atlas": "^0.4"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.15",
         "guzzlehttp/guzzle": "^7.5",
-        "jenky/atlas": "1.x-dev",
-        "jenky/atlas-mock-client": "^1.0",
         "phpstan/phpstan": "^1.10",
-        "phpunit/phpunit": "^9.0",
-        "react/async": "^4.1",
-        "react/http": "^1.9",
-        "symfony/var-dumper": "^6.3"
+        "phpunit/phpunit": "^9.0"
     },
     "autoload": {
         "psr-4": {
@@ -61,10 +57,6 @@
             "@analyse",
             "@test"
         ]
-    },
-    "suggest": {
-        "react/http": "Required to send concurrent requests using React pool",
-        "react/async": "Required to send concurrent requests using React pool"
     },
     "extra": {
         "branch-alias": {
