+++ src/PoolInterface.php
@@ -4,17 +4,22 @@ declare(strict_types=1);
 
 namespace Jenky\Atlas\Pool;
 
-use Jenky\Atlas\Response;
-
+/**
+ * @template TReq
+ * @template TRes
+ */
 interface PoolInterface
 {
+    /**
+     * Set the maximum number of requests to send concurrently.
+     */
     public function concurrent(int $concurrency): PoolInterface;
 
     /**
      * Send concurrent requests.
      *
-     * @param  iterable $requests
-     * @return array<array-key, Response>
+     * @param  iterable<TReq> $requests
+     * @return array<array-key, TRes>
      */
     public function send(iterable $requests): array;
 }
