+++ tests/PoolTest.php
@@ -2,16 +2,31 @@
 
 namespace Jenky\Atlas\Pool\Tests;
 
-use Jenky\Atlas\Pool\Guzzle\Pool as GuzzlePool;
-use Jenky\Atlas\Pool\PoolFactory;
+use Jenky\Atlas\NullConnector;
 use PHPUnit\Framework\TestCase;
 
 class PoolTest extends TestCase
 {
-    public function test_factory(): void
+    public function test_pool(): void
     {
-        $pool = PoolFactory::create(new EchoConnector());
+        $pool = new NullPool(new NullConnector());
 
-        $this->assertInstanceOf(GuzzlePool::class, $pool);
+        $responses = $pool->send([
+            new DummyRequest(),
+            new DummyRequest(),
+        ]);
+
+        $this->assertCount(0, $responses);
+    }
+
+    public function test_concurrent(): void
+    {
+        $pool = new NullPool(new NullConnector());
+
+        $this->expectException(\ValueError::class);
+
+        $pool->concurrent(-1);
+
+        $this->assertNotSame($pool, $pool->concurrent(10), 'Pool is immutable');
     }
 }
